refactor(test): split runtime-import tests out of test_widget_bridge.cpp
Phase 5 P5-1 first cut. test_widget_bridge.cpp is the biggest file in
the entire repo at 14,059 lines — every widget-bridge edit rebuilds
the whole TU, and the file has accumulated 57 section markers covering
~12 distinct compat surfaces (canvas2d, css, yoga, html, runtime-
import, RN, SVG, etc.). The roadmap calls for per-surface splits.

This first cut peels off the most self-contained surface: the 12
TEST_CASE block for `WidgetBridge::install_runtime_import_handlers`
(pulp #468 follow-up), which already lived at the tail of the file
after a "── Runtime design import" section header.

- test_widget_bridge.cpp: 14,059 → 13,609 lines (-450).
- test_widget_bridge_runtime_import.cpp: 479 lines (12 tests, 36
  assertions).
- New CMake target `pulp-test-widget-bridge-runtime-import`
  discovers + runs all 12 cases on its own; the parent target
  `pulp-test-widget-bridge` still discovers + passes the remaining
  392 cases (2257 assertions).
- The small `js_single_quoted` helper (4-case escape switch) is
  duplicated in the new file because the original is `static`
  (file-local) and not exported from any header. A future shared
  `test/test_widget_bridge_helpers.hpp` could centralise it; for now
  the duplication keeps the split behaviour-neutral.

Tests:
- pulp-test-widget-bridge: 384 reported cases (Catch2 counts differ
  from raw TEST_CASE counts when a case has multiple SECTIONs),
  2257 assertions — all pass.
- pulp-test-widget-bridge-runtime-import: 12 cases, 36 assertions —
  all pass.
- Total TEST_CASE count: 392 + 12 = 404 (preserved, was 404 before
  split).

Follow-up cuts will peel canvas2d, css, yoga, RN-style, SVG, and
HTML/ARIA into their own per-surface test TUs.
